+++ Lesson07-Spring-Data/VideoPoint.Lesson07-02.md
@@ -34,6 +34,7 @@ I created module 7 from module 4
 01-25-00 == @Transient   
 01-26-00 == UserRepository   
 01-31-00 == Transactional   
+01-41-00 == Run
