+++ Lesson07-Spring-Data/src/main/java/ru/rrenat358/Lesson07SpringDataApplication.java
@@ -4,10 +4,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class Lesson04SpringBootApplication {
+public class Lesson07SpringDataApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(Lesson04SpringBootApplication.class, args);
+        SpringApplication.run(Lesson07SpringDataApplication.class, args);
     }
 
 }
