+++ Lesson07-Spring-Data/src/main/java/ru/rrenat358/exceptions/EntityNotFoundException.java
@@ -1,0 +1,8 @@
+package ru.rrenat358.exceptions;
+
+public class EntityNotFoundException extends RuntimeException {
+
+    public EntityNotFoundException(String message) {
+        super(message);
+    }
+}
