+++ Lesson07-Spring-Data/src/main/resources/application.properties
@@ -1,5 +1,5 @@
 
-server.port=35813
+server.port=8090
 
 logging.level.ru.rrenat358.contoller=info
