+++ Lesson07-Spring-Data/src/test/java/ru/rrenat358/Lesson07SpringDataApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class Lesson04SpringBootApplicationTests {
+class Lesson07SpringDataApplicationTests {
 
     @Test
     void contextLoads() {
